backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pymongo==4.6.1
motor==3.3.2
pydantic==2.5.2
python-dotenv==1.0.0
openai==1.6.1
requests==2.31.0
Pillow==10.1.0
aiofiles==23.2.1
websockets==12.0
diffusers==0.25.0
transformers==4.36.2
torch==2.1.2
torchvision==0.16.2
accelerate==0.25.0
xformers==0.0.23.post1
compel==2.0.2
invisible-watermark==0.2.0
safetensors==0.4.1
omegaconf==2.3.0
einops==0.7.0
controlnet-aux==0.0.7
opencv-python==4.8.1.78
numpy==1.24.4
scipy==1.11.4